NEW: 	- In-Engine message passing is fully functional. To avoid a little overhead we routed a direct function for in-engine events processing like (Lua) handle_touch_event, when we have a component that needs to be updated when touches happen. 	Example: 		For a simple joypad, it could look like this: 		( Lua Code ) 		event_data = 		{ 			listen_events = 			{ 				-- Since E_TOUCH is defined we must specify a "handle_touch_event function" 				"E_TOUCH" 			},
			handle_event =
			function(this, in_event, in_data)
				-- Handle Custom events there.
			end,

			handle_touch_event =
			function(this, in_x, in_y, in_touchIndex, in_touchId, in_touchType)
				-- Handle engine-specific touch events here.
				this.components["gecJoystick"]:handle_touch(in_x, in_y, in_touchIndex, in_touchId, in_touchType)
			end
		},

_NG

+++ Classes/Engine/Renderer/EAGLView.mm
@@ -150,7 +150,7 @@
 	for ( UITouch* aTouch in touches )
 	{
 		loc = [aTouch locationInView:self];
-		TOUCHABLE_MANAGER->touchesBegan(loc.x, loc.y, aTouch);		
+		TOUCHABLE_MANAGER->touchesBegan(loc.x, loc.y, (int)aTouch);		
 	}
 }
 
@@ -161,7 +161,7 @@
 	for ( UITouch* aTouch in touches )
 	{
 		loc = [aTouch locationInView:self];
-		TOUCHABLE_MANAGER->touchesMoved(loc.x, loc.y, aTouch);		
+		TOUCHABLE_MANAGER->touchesMoved(loc.x, loc.y, (int)aTouch);		
 	}
 }
 
@@ -172,7 +172,7 @@
 	for ( UITouch* aTouch in touches )
 	{
 		loc = [aTouch locationInView:self];
-		TOUCHABLE_MANAGER->touchesEnded(loc.x, loc.y, aTouch);		
+		TOUCHABLE_MANAGER->touchesEnded(loc.x, loc.y, (int)aTouch);		
 	}
 }
